% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{binary.p2skew}
\alias{binary.p2skew}
\title{binary.p2skew}
\usage{
binary.p2skew(p)
}
\arguments{
\item{p}{The binomial probability}
}
\value{
Skew
}
\description{
Converts the probability parameter of a binomial distribution to the skew, assuming n=1.
}
\examples{
 \dontrun{
binary.p2skew(p=.5)
}
}
