% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_interaction_r2_covs.R
\name{generate.interaction.cov.input}
\alias{generate.interaction.cov.input}
\title{Creates the input to \code{power_interaction_r2_covs()}}
\usage{
generate.interaction.cov.input(c.num)
}
\arguments{
\item{c.num}{Number of covariates in the model.}
}
\value{
A list to be used with the 'power_interaction_r2_covs()' function.
}
\description{
Companion function to 'power_interaction_r2_covs()'. Generates a formatted list for users to specify the analysis parameters.
}
\examples{
ex1 = generate.interaction.cov.input(c.num=2)
ex1$correlations$r.y.x1x2 = c(0.1,0.2,0.3)
}
