% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_interaction.R
\name{generate_interaction}
\alias{generate_interaction}
\title{Generate interaction data set}
\usage{
generate_interaction(
  N,
  r.x1.y,
  r.x2.y,
  r.x1x2.y,
  r.x1.x2,
  sd.x1 = 1,
  sd.x2 = 1,
  sd.y = 1,
  mean.x1 = 0,
  mean.x2 = 0,
  mean.y = 0
)
}
\arguments{
\item{N}{Sample size. Must be a positive integer. Has no default value.}

\item{r.x1.y}{Pearson's correlation between x1 and y. Must be between -1 and 1.. Has no default value.}

\item{r.x2.y}{Pearson's correlation between x2 and y. Must be between -1 and 1.. Assumed to be the 'moderator' in some functions. Has no default value.}

\item{r.x1x2.y}{Pearson's correlation between the interaction term x1x2 (x1 * x2) and y. Must be between -1 and 1.. Has no default value.}

\item{r.x1.x2}{Pearson's correlation between x1 and x2. Must be between -1 and 1.. Has no default value.}

\item{sd.x1}{Standard deviation of x1. Defaults to 1.}

\item{sd.x2}{Standard deviation of x2. Defaults to 1.}

\item{sd.y}{Standard deviation of y. Defaults to 1.}

\item{mean.x1}{Mean of x1. Defaults to 0.}

\item{mean.x2}{Mean of x2. Defaults to 0.}

\item{mean.y}{Mean of y. Defaults to 0.}
}
\value{
A data frame containing variables 'x1', 'x2', 'y', and 'x1x2'. 'x1x2' is x1*x2. The correlations between these variables are drawn from the defined population-level values.
}
\description{
Simulate a single data set with an interaction (y ~ x1 + x2 + x1*x2). All values other than 'N' are population-level effects - the values within any single simulated data set will vary around the defined values.
}
\examples{
dataset <- generate_interaction(N = 250,r.x1.y = 0,r.x2.y = .1,r.x1x2.y = -.2,r.x1.x2 = .3)
cor(dataset)
}
