% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_key.R
\docType{data}
\name{name_key}
\alias{name_key}
\title{Name key for plotting}
\format{
A data frame with 15 rows and 2 variables
}
\usage{
data(name_key)
}
\description{
Expanded variable names so that plots look nicer.
}
