% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{norm2binary}
\alias{norm2binary}
\title{norm2binary}
\usage{
norm2binary(x, skew)
}
\arguments{
\item{x}{Input vector}

\item{skew}{Desired output skew}
}
\value{
A binary variable
}
\description{
Transforms a vector with a normal distribution to a binomial distribution with two values.
}
\examples{
norm2binary(x = rnorm(n = 100,mean = 0,sd = 1), skew = 1)
}
