% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{norm2gamma}
\alias{norm2gamma}
\title{norm2gamma}
\usage{
norm2gamma(x, skew)
}
\arguments{
\item{x}{Input vector}

\item{skew}{Desired skew}
}
\value{
A vector with a (skewed) gamma distribution
}
\description{
Transforms a vector with a normal distribution to a gamma distribution.
}
\examples{
norm2gamma(x = rnorm(n = 100,mean = 0,sd = 1), skew = 1)
}
