% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{norm2likert}
\alias{norm2likert}
\title{norm2likert}
\usage{
norm2likert(x, skew, k)
}
\arguments{
\item{x}{Input vector}

\item{skew}{Desired output skew}

\item{k}{Number of discrete values (e.g., 2=binary, 5=likert scale)}
}
\value{
A likert or binary variable
}
\description{
Transforms a vector with a normal distribution to a binomial distribution with two values.
}
\examples{
\dontrun{
norm2likert(x = rnorm(n = 100,mean = 0,sd = 1), skew = 1,k=2)
}
}
