% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{norm2ordinal}
\alias{norm2ordinal}
\title{norm2ordinal}
\usage{
norm2ordinal(x, k)
}
\arguments{
\item{x}{Input vector}

\item{k}{Number of discrete values (e.g., 2=binary, 5=likert scale)}
}
\value{
A ordinal or binary variable
}
\description{
Transforms a vector with a normal distribution to a binomial distribution with two values.
}
\examples{
norm2ordinal(x = rnorm(n = 100,mean = 0,sd = 1),k=2)
}
