% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction.R
\name{plot_interaction}
\alias{plot_interaction}
\title{Plot interaction}
\usage{
plot_interaction(data, q = 3)
}
\arguments{
\item{data}{Output of generate_interaction().}

\item{q}{Simple slope quantiles. Default is 2. X2 is the default moderator, unless X1 is already binary. Must be a positive integer > 1.}
}
\value{
A ggplot2 object
}
\description{
Plots a single simulated interaction data set
}
\examples{
dataset <- generate_interaction(N = 250,r.x1.y = 0,r.x2.y = .1,r.x1x2.y = -.2,r.x1.x2 = .3)
plot_interaction(dataset,q=3)
}
