% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_power_curve}
\alias{plot_power_curve}
\title{Plot the results of find_sample_size() as one line per effect}
\usage{
plot_power_curve(results, desired_power = 0.8)
}
\arguments{
\item{results}{The dataframe returned by find_sample_size().}

\item{desired_power}{Power level to mark with a dashed line.}
}
\value{
A ggplot.
}
\description{
Plot the results of find_sample_size() as one line per effect
}
