% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{power_interaction}
\alias{power_interaction}
\title{Power of x, z and their interaction at one sample size}
\usage{
power_interaction(n, b_x, b_z, b_interaction, n_simulations = 1000)
}
\arguments{
\item{n}{Sample size.}

\item{b_x}{Standardized effect of x.}

\item{b_z}{Standardized effect of z.}

\item{b_interaction}{Standardized effect of the x by z interaction.}

\item{n_simulations}{Number of simulated datasets.}
}
\value{
Three powers, in the order x, z, interaction.
}
\description{
Power of x, z and their interaction at one sample size
}
