% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_interaction_data}
\alias{simulate_interaction_data}
\title{Simulate one dataset with an interaction effect}
\usage{
simulate_interaction_data(n, b_x, b_z, b_interaction)
}
\arguments{
\item{n}{Sample size.}

\item{b_x}{Standardized effect of x.}

\item{b_z}{Standardized effect of z.}

\item{b_interaction}{Standardized effect of the x by z interaction.}
}
\value{
A dataframe with the columns x, z and y.
}
\description{
Simulate one dataset with an interaction effect
}
