% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_interaction.R
\name{power_interaction}
\alias{power_interaction}
\title{Power analysis for interactions}
\usage{
power_interaction(
  n.iter,
  N,
  r.x1.y,
  r.x2.y,
  r.x1x2.y,
  r.x1.x2,
  sd.x1 = 1,
  sd.x2 = 1,
  sd.y = 1,
  mean.x1 = 0,
  mean.x2 = 0,
  mean.y = 0,
  alpha = 0.05,
  q = 2,
  cl = NULL,
  ss.IQR = 1.5,
  detailed_results = FALSE,
  full_simulation = FALSE
)
}
\arguments{
\item{n.iter}{Number of iterations. The number of simulations to run for each unique setting combination. Must be a positive integer.}

\item{N}{Sample size. Must be a positive integer. Has no default value. Can be a single value or a vector of values.}

\item{r.x1.y}{Pearson's correlation between x1 and y. Must be between -1 and 1.. Has no default value. Can be a single value or a vector of values.}

\item{r.x2.y}{Pearson's correlation between x2 and y. Must be between -1 and 1.. Assumed to be the 'moderator' in some functions. Has no default value. Can be a single value or a vector of values.}

\item{r.x1x2.y}{Pearson's correlation between the interaction term x1x2 (x1 * x2) and y. Must be between -1 and 1.. Has no default value. Can be a single value or a vector of values.}

\item{r.x1.x2}{Pearson's correlation between x1 and x2. Must be between -1 and 1.. Has no default value. Can be a single value or a vector of values.}

\item{sd.x1}{Standard deviation of x1. Defaults to 1. Can be a single value or a vector of values.}

\item{sd.x2}{Standard deviation of x2. Defaults to 1. Can be a single value or a vector of values.}

\item{sd.y}{Standard deviation of y. Defaults to 1. Can be a single value or a vector of values.}

\item{mean.x1}{Mean of x1. Defaults to 0. Can be a single value or a vector of values.}

\item{mean.x2}{Mean of x2. Defaults to 0. Can be a single value or a vector of values.}

\item{mean.y}{Mean of y. Defaults to 0. Can be a single value or a vector of values.}

\item{alpha}{The alpha. At what p-value is the interaction deemed significant? Default is 0.05.}

\item{q}{Simple slopes. How many quantiles should x2 be split into for simple slope testing? Default is 2. Simple slope testing returns the effect-size (slope) of y~x1 for the two most extreme quantiles of x2. If q=3 then the two slopes are y~x1 for the bottom 33\% of x2, and the top 33\% of x2.}

\item{cl}{Number of clusters to use for running simulations in parallel (recommended). Default is 1 (i.e. not in parallel).}

\item{ss.IQR}{Simple slope IQR. Multiplier when estimating the distribution of simple slopes within each simulation setting. Default is 1.5.}

\item{detailed_results}{Default is FALSE. Should detailed results be reported (mean of effect size and correlations for each setting)?}

\item{full_simulation}{Default is FALSE. If TRUE, will return a list that includes the full per-simulation results.}
}
\value{
A data frame containing the power (\% significant results) for each unique setting combination. If full_simulation = TRUE will return a list, with one data frame that includes power, and a second that includes raw simulation results.
}
\description{
Power analysis for interaction models, by simulation. A set of n.iter simulations is run for each unique combination of model settings.
}
\examples{
\dontrun{
power_interaction(n.iter=1000, N=seq(100,300,by=10),r.x1.y=0.2, r.x2.y=.2,r.x1x2.y=0.5,r.x1.x2=.2)
}

}
